+++ setup_kvm.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+curl -sSL https://raw.githubusercontent.com/aspnet/Home/dev/kvminstall.sh | sh && source ~/.kre/kvm/kvm.sh && kvm upgrade