+++ ui/lib/hapi/underscore_compiler.js
@@ -1,7 +1,0 @@
-module.exports = function(templ, options) {
-
-  return function(ctx, options) {
-    return require('underscore').template(templ)(ctx || null);
-  };
-
-};